[UPD/WIP] change reset logic of the team registration
