+++ _data/handles/0xsi.json
@@ -1,0 +1,5 @@
+{
+  "handle": "0xsi",
+  "link": "",
+  "moralisId": "AxfKfPo2eD0SiB5L6OAG5fbT"
+}
